Implement a ChainedVector array type for chaining arrays together as one
This was another piece of functionality that was slightly half-baked,
not fully tested in CSV.jl. When multi-threaded parsing is used in
CSV.jl, each thread produces a column "chunk" which get "chained"
together after parsing. This implementation includes a field for the
chained arrays, as well as a field for the last "indexes" for each
chunk. This allows for efficient use of `searchsortedfirst`, a binary
search function in Base to index individual elements.

We technically allow chaining any `AbstractVector`, but mutable
operations will require the chained array type to support mutable
operations as well; e.g. to `insert!` into a ChainedVector, we need to
be able to `insert!` into the underlying array.