Fix a rare corner case when cycling a sentinel value
sudete on discourse pointed out that there's a corner case not currently
covered when cycling to a new sentinel value. We might try to be
`setindex!`ing a value that is equal to our sentinel value, so we go to
pick a new sentinel, and it might be possible for the newly chosen
sentinel to be equal to the current sentinel, which would pass the "new
sentinel" check because it doesn't exist in the current array. The fix
here is to ensure the newly chosen sentinel is also not the same as the
existing sentinel