Fix deleteat! with empty inds array
Fixes https://github.com/JuliaData/CSV.jl/issues/690. The issue here is we weren't
testing or accounting for the case when the `inds` arguement to deleteat! is empty.
This comes up in practice from `dropmissing!` in DataFrames.jl when there are no missing
values in a column, it still calls `deleteat!` with the empty array.
The fix is pretty straightforward: we can just remove the bounds checks here because
we're doing the bounds check in the underlying `::Integer` method anyway.