Make deletat! for SentinelArray/ChainedVector much faster
Previously, we were taking the most naive route possible. These new
algorithms are now much faster and attempt to take as many fast paths as
possible. For SentinelArray, we use the Base definition for all cases
except when a SentinelArray uses `#undef` for sentinel and has multiple
deletions, since the Base algorithm is inconsistent in allowing `#undef`
to be deleted. We copy-paste the Base algorithm in that case and handle
the `#undef`s specially.

For ChainedVector, we use a new algorithm that iterates indices to
delete and accumulates them per "chunk", thus moving from an algorithm
of one deletion per index, to one deletion call per chunk, which makes
it must faster.

Fixed #23.