Fix iteration issue when chains have zero-length
Fixes #27. The issue here is that `iterate` for `ChainedVector` was
assuming that underlying array chains wouldn't have zero-length. If a
user is doing a lot of filtering/deleting, however, it might be the case
that certain chunks end up with zero-length (though we do try to prune
those out when possible, so I'm still a little unsure how we get in this
state). Nevertheless, this assumption is a bit optimistic, and we can do
better by just checking if the next chunk is zero-length when iterating
and moving on to the next chunk if so.