Optimize SentinelArray wrapping a ChainedVector
Fixes #30. The issue here is that we were missing an optimized
implementation of `isassigned` for ChainedVector, so when SentinelArray
was calling `getindex` with a reference type that used `#undef` as
sentinel, we were falling back to the generic Base definition of
`isassigned` for all `AbstractArray` that uses `try-catch` and is
dreadfully slow. Just adding a pass-through definition of `isassigned`
brings the performance back to where you would expect.