Fix append on Vector with ChainedVector (#48)
Fixes issue #45. The issue here is that the `Base.append!`
implementation relies on calling `first(eachindex(items))`, which has
now been optimized to return a custom type for `ChainedVector`. The
quick work-around is we just overload `Base.append!` when the appending
items is `ChainedVector`, since the definition is so small anyway. If we
end up hearing of other issues with this custom `eachindex` on
`ChainedVector`, we may need to move it back to a more traditional
definition.