Define has for ChainedVectorIndex
Fixes #62. It seems some code path in plotting is storing elements of
`eachindex` in a `Dict`, which is resulting in `hash` being called on
our custom index type `ChainedVectorIndex`. As a reminder, Our
`ChainedVectorIndex` is _basically_ a plain Integer, but we just keep
track of the current array/array index of the individual chunk we're
iterating which makes iteration overall on `ChainedVector`s much more
efficient. So, the fix here is pretty simply in that we just define
`hash` to forward to the underlying index.